Stop doing a stupid division

+++ firmware/src/battery_mon.h
@@ -2,7 +2,7 @@
 
 #include "adc12_a.h"
 
-extern volatile uint16_t batt_voltage;
+extern volatile uint8_t batt_voltage;
 extern ADC12_A_configureMemoryParam ADC_config;
 
 extern void setup_bat_monitor(void);
